+++ packages/accessibility/src/behaviors/Toolbar/toolbarBehavior.ts
@@ -1,5 +1,5 @@
 import { Accessibility } from '../../types'
-import { FocusZoneMode, FocusZoneDirection } from '../../focusZone/types'
+import { FocusZoneDirection } from '../../focusZone/types'
 
 /**
  * @description
@@ -7,18 +7,16 @@ import { FocusZoneMode, FocusZoneDirection } from '../../focusZone/types'
  * Child item components need to have toolbarItemBehavior assigned.
  * @specification
  * Adds role 'toolbar' to 'root' slot.
- * Embeds component into FocusZone.
  * Provides arrow key navigation in horizontal direction.
  * When component's container element receives focus, focus will be set to the default focusable child element of the component.
  */
-const toolbarBehavior: Accessibility = (props: any) => ({
+const toolbarBehavior: Accessibility = props => ({
   attributes: {
     root: {
       role: 'toolbar',
     },
   },
   focusZone: {
-    mode: FocusZoneMode.Embed,
     props: {
       shouldFocusInnerElementWhenReceivedFocus: true,
       direction: FocusZoneDirection.horizontal,